Add files via upload
Oasis_Infobyte_Internship
TASK 2: NUMBER GUESSING GAME QUESTION: The fun and easy project "Guess the Number" is a short Java project that allows the user to guess the number generated by the computer & involves the following steps:

The system generates a random number from a given range, say 1 to 100.
The user is prompted to enter their given number in a displayed dialogue box.
The computer then tells if the entered number matches the guesses number or it is higher/lower than the generated number.
The game continues under the user guessing the number. You can also incorporate further details as:
Limiting the number of attempts. Adding more rounds. Displaying score. Giving points based on the number of attempts.

TASK 3:ATM INTERFACE

We have all come across ATMs in our cities and it is built on Java. This complex project consists of
five different classes and is a console-based application. When the system starts the user is prompted with user id and user pin. On entering the details successfully, then ATM functionalities are unlocked. The project allows to perform following operations:

1)Transactions History

2)Withdraw

3)Deposit

4)Transfer

5)Quit

For Demo Click

TASK 4: NUMBER GUESSING GAME QUESTION: The project should be capable of the following functionalities.

Login
Update Profile and Password
Selecting answers for MCQs
Timer and auto submit
Closing session and Logout